Feedback Implementation.
Multiple UI Prefabs. Enemy Feedback (Faster, own NavMesh Agent [Problems with GIGA Ranged Robot], Pasu-Kan RageMode instead of Jump, MultiShot for RangedRobot, AttackSpeed Evaluation, HealthComponent Bug Fix Attempt).

+++ Assets/Lucas/Scripts/Stats/BasicEnemy/BasicEnemy/BasicEnemyData.cs
@@ -8,7 +8,7 @@ public class BasicEnemyData : ScriptableObject
     public int _damagePerHit = 100;
     public float _attackSpeed = 1;
     public int _givenXP = 1;
-    public int _moveSpeed = 1;
+    public float _moveSpeed = 1;
     public int _armor = 1;
     public int _elementalResistance = 1;
     public int _technicalResistance = 1;
